+++ crates/cheatnet/src/state.rs
@@ -1,12 +1,14 @@
 use crate::forking::state::ForkStateReader;
 use crate::runtime_extensions::call_to_blockifier_runtime_extension::rpc::{
-    subtract_execution_resources, CallResult,
+    subtract_execution_resources, AddressOrClassHash, CallResult,
 };
 use crate::runtime_extensions::forge_runtime_extension::cheatcodes::spoof::TxInfoMock;
 use crate::runtime_extensions::forge_runtime_extension::cheatcodes::spy_events::{
     Event, SpyTarget,
 };
-use blockifier::execution::entry_point::{CallEntryPoint, ExecutionResources};
+use blockifier::execution::entry_point::{
+    CallEntryPoint, EntryPointExecutionResult, ExecutionResources,
+};
 use blockifier::{
     execution::contract_class::ContractClass,
     state::state_api::{StateReader, StateResult},
@@ -18,18 +20,20 @@ use runtime::starknet::state::DictStateReader;
 use starknet_api::core::EntryPointSelector;
 
 use crate::constants::{build_test_entry_point, TEST_CONTRACT_CLASS_HASH};
+use blockifier::execution::call_info::CallInfo;
 use blockifier::state::errors::StateError::UndeclaredClassHash;
 use starknet_api::transaction::ContractAddressSalt;
 use starknet_api::{
+    class_hash,
     core::{ClassHash, CompiledClassHash, ContractAddress, Nonce},
-    hash::StarkFelt,
-    stark_felt,
+    hash::{StarkFelt, StarkHash},
     state::StorageKey,
 };
 use std::cell::{Ref, RefCell};
 use std::collections::HashMap;
 use std::hash::BuildHasher;
 use std::rc::Rc;
+use trace_data::L1Resources;
 
 // Specifies which contracts to target
 // with a cheatcode function
@@ -40,7 +44,7 @@ pub enum CheatTarget {
 }
 
 // Specifies the duration of the cheat
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub enum CheatSpan {
     Indefinite,
     Number(usize),
@@ -130,6 +134,7 @@ impl StateReader for ExtendedStateReader {
     }
 }
 
+#[derive(Debug)]
 pub enum CheatStatus<T> {
     Cheated(T, CheatSpan),
     Uncheated,
@@ -147,16 +152,17 @@ impl<T> CheatStatus<T> {
 }
 
 /// Tree structure representing trace of a call.
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct CallTrace {
     pub entry_point: CallEntryPoint,
     // These also include resources used by internal calls
     pub used_execution_resources: ExecutionResources,
+    pub used_l1_resources: L1Resources,
     pub nested_calls: Vec<Rc<RefCell<CallTrace>>>,
     pub result: CallResult,
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 struct CallStackElement {
     // when we exit the call we use it to calculate resources used by the call
     resources_used_before_call: ExecutionResources,
@@ -164,6 +170,7 @@ struct CallStackElement {
     cheated_data: CheatedData,
 }
 
+#[derive(Debug)]
 pub struct NotEmptyCallStack(Vec<CallStackElement>);
 
 impl NotEmptyCallStack {
@@ -194,10 +201,8 @@ impl NotEmptyCallStack {
     }
 
     pub fn top_cheated_data(&mut self) -> CheatedData {
-        let top_val = self.0.pop().unwrap();
-        let borrowed_ref = top_val.cheated_data.clone();
-        self.0.push(top_val);
-        borrowed_ref
+        let top_val = self.0.last().unwrap();
+        top_val.cheated_data.clone()
     }
 
     fn pop(&mut self) -> CallStackElement {
@@ -225,10 +230,12 @@ pub struct CheatedData {
     pub tx_info: Option<TxInfoMock>,
 }
 
+#[derive(Debug)]
 pub struct TraceData {
     pub current_call_stack: NotEmptyCallStack,
 }
 
+#[derive(Debug)]
 pub struct CheatnetState {
     pub rolled_contracts: HashMap<ContractAddress, CheatStatus<Felt252>>,
     pub global_roll: Option<(Felt252, CheatSpan)>,
@@ -251,10 +258,11 @@ pub struct CheatnetState {
 impl Default for CheatnetState {
     fn default() -> Self {
         let mut test_code_entry_point = build_test_entry_point();
-        test_code_entry_point.class_hash = Some(ClassHash(stark_felt!(TEST_CONTRACT_CLASS_HASH)));
+        test_code_entry_point.class_hash = Some(class_hash!(TEST_CONTRACT_CLASS_HASH));
         let test_call = Rc::new(RefCell::new(CallTrace {
             entry_point: test_code_entry_point,
             used_execution_resources: Default::default(),
+            used_l1_resources: Default::default(),
             nested_calls: vec![],
             result: CallResult::Success { ret_data: vec![] },
         }));
@@ -387,6 +395,7 @@ impl TraceData {
         let new_call = Rc::new(RefCell::new(CallTrace {
             entry_point,
             used_execution_resources: Default::default(),
+            used_l1_resources: Default::default(),
             nested_calls: vec![],
             result: CallResult::Success { ret_data: vec![] },
         }));
@@ -409,7 +418,8 @@ impl TraceData {
     pub fn exit_nested_call(
         &mut self,
         resources_used_after_call: &ExecutionResources,
-        call_result: CallResult,
+        execution_result: &EntryPointExecutionResult<CallInfo>,
+        identifier: &AddressOrClassHash,
     ) {
         let CallStackElement {
             resources_used_before_call,
@@ -420,7 +430,19 @@ impl TraceData {
         let mut last_call = last_call.borrow_mut();
         last_call.used_execution_resources =
             subtract_execution_resources(resources_used_after_call, &resources_used_before_call);
-        last_call.result = call_result;
+
+        last_call.used_l1_resources.l2_l1_message_sizes = execution_result.as_ref().map_or_else(
+            |_| vec![],
+            |call_info| {
+                let messages = &call_info.execution.l2_to_l1_messages;
+                messages
+                    .iter()
+                    .map(|ordered_message| ordered_message.message.payload.0.len())
+                    .collect()
+            },
+        );
+
+        last_call.result = CallResult::from_execution_result(execution_result, identifier);
     }
 }