+++ crates/cheatnet/tests/cheatcodes/mock_call.rs
@@ -8,7 +8,6 @@ use crate::{
 use cairo_felt::Felt252;
 use cheatnet::runtime_extensions::forge_runtime_extension::cheatcodes::declare::declare;
 use cheatnet::state::CheatnetState;
-use conversions::felt252::FromShortString;
 use conversions::IntoConv;
 use starknet_api::core::ContractAddress;
 
@@ -18,33 +17,31 @@ trait MockCallTrait {
     fn start_mock_call(
         &mut self,
         contract_address: &ContractAddress,
-        function_name: &str,
+        function_selector: Felt252,
         ret_data: &[u128],
     );
-    fn stop_mock_call(&mut self, contract_address: &ContractAddress, function_name: &str);
+    fn stop_mock_call(&mut self, contract_address: &ContractAddress, function_selector: Felt252);
 }
 
 impl<'a> MockCallTrait for TestEnvironment<'a> {
     fn start_mock_call(
         &mut self,
         contract_address: &ContractAddress,
-        function_name: &str,
+        function_selector: Felt252,
         ret_data: &[u128],
     ) {
-        let function_name = Felt252::from_short_string(function_name).unwrap();
         let ret_data: Vec<Felt252> = ret_data.iter().map(|x| Felt252::from(*x)).collect();
         self.runtime_state.cheatnet_state.start_mock_call(
             *contract_address,
-            &function_name,
+            function_selector,
             &ret_data,
         );
     }
 
-    fn stop_mock_call(&mut self, contract_address: &ContractAddress, function_name: &str) {
-        let function_name = Felt252::from_short_string(function_name).unwrap();
+    fn stop_mock_call(&mut self, contract_address: &ContractAddress, function_selector: Felt252) {
         self.runtime_state
             .cheatnet_state
-            .stop_mock_call(*contract_address, &function_name);
+            .stop_mock_call(*contract_address, function_selector);
     }
 }
 
@@ -66,7 +63,7 @@ fn mock_call_simple() {
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -99,7 +96,7 @@ fn mock_call_stop() {
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -113,10 +110,9 @@ fn mock_call_stop() {
 
     assert_success!(output, ret_data);
 
-    runtime_state.cheatnet_state.stop_mock_call(
-        contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
-    );
+    runtime_state
+        .cheatnet_state
+        .stop_mock_call(contract_address, felt_selector_from_name("get_thing"));
 
     let output = call_contract(
         &mut cached_state,
@@ -144,10 +140,9 @@ fn mock_call_stop_no_start() {
 
     let selector = felt_selector_from_name("get_thing");
 
-    runtime_state.cheatnet_state.stop_mock_call(
-        contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
-    );
+    runtime_state
+        .cheatnet_state
+        .stop_mock_call(contract_address, felt_selector_from_name("get_thing"));
 
     let output = call_contract(
         &mut cached_state,
@@ -176,18 +171,14 @@ fn mock_call_double() {
     let selector = felt_selector_from_name("get_thing");
 
     let ret_data = vec![Felt252::from(123)];
-    runtime_state.cheatnet_state.start_mock_call(
-        contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
-        &ret_data,
-    );
+    runtime_state
+        .cheatnet_state
+        .start_mock_call(contract_address, selector.clone(), &ret_data);
 
     let ret_data = vec![Felt252::from(999)];
-    runtime_state.cheatnet_state.start_mock_call(
-        contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
-        &ret_data,
-    );
+    runtime_state
+        .cheatnet_state
+        .start_mock_call(contract_address, selector.clone(), &ret_data);
 
     let output = call_contract(
         &mut cached_state,
@@ -199,10 +190,9 @@ fn mock_call_double() {
 
     assert_success!(output, ret_data);
 
-    runtime_state.cheatnet_state.stop_mock_call(
-        contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
-    );
+    runtime_state
+        .cheatnet_state
+        .stop_mock_call(contract_address, selector.clone());
 
     let output = call_contract(
         &mut cached_state,
@@ -233,7 +223,7 @@ fn mock_call_double_call() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -275,7 +265,7 @@ fn mock_call_proxy() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -324,7 +314,7 @@ fn mock_call_proxy_with_other_syscall() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -374,7 +364,7 @@ fn mock_call_inner_call_no_effect() {
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -428,7 +418,7 @@ fn mock_call_library_call_no_effect() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_constant_thing").unwrap(),
+        felt_selector_from_name("get_constant_thing"),
         &ret_data,
     );
 
@@ -461,7 +451,7 @@ fn mock_call_before_deployment() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         precalculated_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
@@ -504,7 +494,7 @@ fn mock_call_not_implemented() {
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing_not_implemented").unwrap(),
+        felt_selector_from_name("get_thing_not_implemented"),
         &ret_data,
     );
 
@@ -533,7 +523,7 @@ fn mock_call_in_constructor() {
     let ret_data = vec![Felt252::from(223)];
     runtime_state.cheatnet_state.start_mock_call(
         balance_contract_address,
-        &Felt252::from_short_string("get_balance").unwrap(),
+        felt_selector_from_name("get_balance"),
         &ret_data,
     );
 
@@ -579,13 +569,13 @@ fn mock_call_two_methods() {
     let ret_data = vec![Felt252::from(123)];
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_constant_thing").unwrap(),
+        felt_selector_from_name("get_constant_thing"),
         &ret_data,
     );
 
@@ -623,7 +613,7 @@ fn mock_call_nonexisting_contract() {
 
     runtime_state.cheatnet_state.start_mock_call(
         contract_address,
-        &Felt252::from_short_string("get_thing").unwrap(),
+        felt_selector_from_name("get_thing"),
         &ret_data,
     );