+++ crates/runtime/src/starknet/context.rs
@@ -18,17 +18,16 @@ use cairo_vm::vm::runners::builtin_runner::{
 };
 use cairo_vm::vm::runners::cairo_runner::RunResources;
 use serde::{Deserialize, Serialize};
-use starknet_api::contract_address;
-use starknet_api::data_availability::DataAvailabilityMode;
 
 use starknet_api::block::{BlockNumber, BlockTimestamp};
-use starknet_api::transaction::{Resource, ResourceBounds, ResourceBoundsMapping};
-use starknet_api::{
-    core::{ChainId, ContractAddress, Nonce, PatriciaKey},
-    hash::{StarkFelt, StarkHash},
-    patricia_key,
-    transaction::{TransactionHash, TransactionSignature, TransactionVersion},
+use starknet_api::core::{ChainId, ContractAddress, Nonce, PatriciaKey};
+use starknet_api::data_availability::DataAvailabilityMode;
+use starknet_api::hash::{StarkFelt, StarkHash};
+use starknet_api::transaction::{
+    Resource, ResourceBounds, ResourceBoundsMapping, TransactionHash, TransactionSignature,
+    TransactionVersion,
 };
+use starknet_api::{contract_address, patricia_key};
 
 pub const DEFAULT_BLOCK_NUMBER: u64 = 2000;
 pub const SEQUENCER_ADDRESS: &str = "0x1000";
@@ -163,7 +162,7 @@ impl Default for BlockInfo {
         Self {
             block_number: BlockNumber(DEFAULT_BLOCK_NUMBER),
             timestamp: BlockTimestamp::default(),
-            sequencer_address: ContractAddress(patricia_key!(SEQUENCER_ADDRESS)),
+            sequencer_address: contract_address!(SEQUENCER_ADDRESS),
         }
     }
 }